feat(graphql): query purchase history for an nft
