+++ crates/graphql/src/schema/context.rs
@@ -4,15 +4,15 @@ use objects::{
     bid_receipt::BidReceipt,
     listing::{Bid, Listing},
     listing_receipt::ListingReceipt,
-    purchase_receipt::PurchaseReceipt,
     nft::{Nft, NftAttribute, NftCreator, NftOwner},
+    purchase_receipt::PurchaseReceipt,
     stats::{MarketStats, MintStats},
     store_creator::StoreCreator,
     storefront::Storefront,
 };
 use scalars::{markers::StoreConfig, PublicKey};
 
-use super::{prelude::*};
+use super::prelude::*;
 
 #[derive(Clone)]
 pub struct AppContext {
