+++ crates/graphql/src/schema/dataloaders/mod.rs
@@ -2,14 +2,12 @@ pub mod auction_house;
 pub mod bid_receipt;
 pub mod collection;
 pub mod listing;
-pub mod listing_receipt;
 pub mod nft;
 pub mod stats;
 pub mod store_creator;
 pub mod storefront;
 
 pub(self) mod batcher;
-pub mod purchase_receipt;
 
 pub(self) mod prelude {
     pub use async_trait::async_trait;
