+++ crates/graphql/src/schema/dataloaders/nft.rs
@@ -1,6 +1,12 @@
-use objects::nft::{Nft, NftAttribute, NftCreator, NftOwner};
+use objects::{
+    listing_receipt::ListingReceipt,
+    nft::{Nft, NftAttribute, NftCreator, NftOwner},
+    purchase_receipt::PurchaseReceipt,
+};
 use scalars::PublicKey;
-use tables::{attributes, metadata_creators, token_accounts};
+use tables::{
+    attributes, listing_receipts, metadata_creators, metadatas, purchase_receipts, token_accounts,
+};
 
 use super::prelude::*;
 
@@ -74,5 +80,56 @@ impl TryBatchFn<PublicKey<Nft>, Option<NftOwner>> for Batcher {
                 })
             })
             .batch(mint_addresses))
+    }
+}
+
+#[async_trait]
+impl TryBatchFn<PublicKey<Nft>, Vec<PurchaseReceipt>> for Batcher {
+    async fn load(
+        &mut self,
+        addresses: &[PublicKey<Nft>],
+    ) -> TryBatchMap<PublicKey<Nft>, Vec<PurchaseReceipt>> {
+        let conn = self.db()?;
+
+        let rows: Vec<models::PurchaseReceipt> = purchase_receipts::table
+            .inner_join(metadatas::table.on(metadatas::address.eq(purchase_receipts::metadata)))
+            .select(purchase_receipts::all_columns)
+            .filter(purchase_receipts::metadata.eq(any(addresses)))
+            .order(purchase_receipts::created_at.desc())
+            .load(&conn)
+            .context("Failed to load purchase receipts")?;
+
+        Ok(rows
+            .into_iter()
+            .map(|purchase| (purchase.metadata.clone(), purchase.try_into()))
+            .batch(addresses))
+    }
+}
+
+#[async_trait]
+impl TryBatchFn<PublicKey<Nft>, Vec<ListingReceipt>> for Batcher {
+    async fn load(
+        &mut self,
+        addresses: &[PublicKey<Nft>],
+    ) -> TryBatchMap<PublicKey<Nft>, Vec<ListingReceipt>> {
+        let conn = self.db()?;
+
+        let rows: Vec<models::ListingReceipt> = listing_receipts::table
+            .inner_join(metadatas::table.on(metadatas::address.eq(listing_receipts::metadata)))
+            .inner_join(
+                token_accounts::table.on(token_accounts::mint_address.eq(metadatas::mint_address)),
+            )
+            .select(listing_receipts::all_columns)
+            .filter(token_accounts::amount.eq(1))
+            .filter(listing_receipts::canceled_at.is_null())
+            .filter(listing_receipts::purchase_receipt.is_null())
+            .filter(listing_receipts::metadata.eq(any(addresses)))
+            .load(&conn)
+            .context("Failed to load listing receipts")?;
+
+        Ok(rows
+            .into_iter()
+            .map(|listing| (listing.metadata.clone(), listing.try_into()))
+            .batch(addresses))
     }
 }
